requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
flake8==3.6.0
google-api-core==1.6.0
google-auth==1.6.1
google-cloud-automl==0.1.1
googleapis-common-protos==1.5.5
grpcio==1.17.1
idna==2.8
mccabe==0.6.1
pbr==4.1.0
pep8==1.7.1
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.7
requests==2.21.0
rsa==4.0
six==1.11.0
stevedore==1.28.0
urllib3==1.24.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
